Have validation in examples/form_table.rb complain about nil values (no value entered) / Fix Reset button in examples/meta_example.rb when on a version other than version 1

+++ lib/glimmer/libui/attributed_string.rb
@@ -197,6 +197,7 @@ module Glimmer
       end
       
       def destroy
+        return if ControlProxy.main_window_proxy&.destroying?
         open_type_features.destroy unless open_type_features.nil?
         @parent_proxy&.children&.delete(self)
       end
